backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-bcrypt==1.0.1
flask-cors==5.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
marshmallow==3.25.1
python-dotenv==1.1.0
